src/goblin/requirements.txt
click>=8.1
httpx>=0.27
pyyaml>=6.0